requirements.txt
clip==1.0
cycler==0.10.0
ftfy==5.9
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2021.3.17
six==1.15.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.59.0
typing-extensions==3.7.4.3
wcwidth==0.2.5
